Fix camera rotating issue + implement scrolling
All I had to do was add a program to "frame" instead of "screen" at the last line of the constructor. By the way, I implemented a "renderToTexture" static variable that toggles render-to-texture for the entire camera rendering system, which adds an extra display, buffer, and  as the framebuffer, so I'll be doing nerdy math for rotating the camera if render-to-texture is off by manipulating all the sprites' pivot using formula.