Major state switching and pause screen revamp
Implementing logic for hiding the rating popup and ui required moving the countdown display and the pause screen to the playfield because the if you hide the hud it won't be initialized. So this means if the countdown display and hud were still in there, they would simply not show up.

This also implements a fancier way to switch states via enums. Now, time for me to cook up the real main menu assets.

This change again also moves the pause screen to its own buffer and program. This means more stuff can be added there since you don't have to worry about the overhead of fragment shader code injections.

Oh yeah and there's an extra minor change being that the note spritesheet has been extruded to fix a bit of texture bleeding.