Options menu BETA: barebones
Keybinds aren't selectable yet and everything else such as the preferences menu and graphics menu have not been created yet. This change also implements the options menu on the pause menu.

I need serious help on it because I fucking hate working on it already and now when the game pauses it just fucking restarts