Replace mouse moving with mouse wheeling entirely
The titlebar not appearing was all because of another minor coding oversight

fucking incredible

Anyway I just made it so that the app screen is actually inside the window like how operating systems fit their app screen in their window.

I also cut off mouse control support because it's useless since now we have a custom window and the only things we need mouse support of is for the freeplay menu, the pause screen. and the options menu. Anyway I replaced the mouse control support on the main menu with mouse wheel
support cause it's at least fucking better.